requirements.txt
# ...existing code...

google-auth-oauthlib==1.2.0
google-auth==2.29.0
gspread==6.1.0
pymongo[srv]>=4.6.0
python-dotenv>=1.0.0
dnspython>=2.4.0
